Backend/exam_proctoring/requirements.txt
asgiref==3.8.1
channels==4.2.2
channels_redis==4.2.1
dataclasses==0.6
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djongo==1.2.31
dnspython==2.7.0
msgpack==1.1.0
pymongo==4.13.0
python-decouple==3.8
pytz==2025.2
redis==6.2.0
sqlparse==0.5.3
